Refactored app layer. Added unitofwork. Added some repos.

+++ Dddreams.Application/Features/Dreams/Commands/Delete/DeleteDreamCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using Dddreams.Application.Common.Exceptions;
 using Dddreams.Application.Helpers;
 using Dddreams.Application.Interfaces.Repositories;
+using Dddreams.Domain.Enums;
 using MediatR;
 
 namespace Dddreams.Application.Features.Dreams.Commands.Delete;
@@ -10,12 +11,13 @@ public class DeleteDreamCommandHandler : IRequestHandler<DeleteDreamCommand, boo
 
     private readonly IDreamsRepository _dreamsRepository;
     private readonly IUserRepository _userRepository;
+    private readonly IUnitOfWork _unitOfWork;
 
-
-    public DeleteDreamCommandHandler(IDreamsRepository dreamsRepository, IUserRepository userRepository)
+    public DeleteDreamCommandHandler(IDreamsRepository dreamsRepository, IUserRepository userRepository, IUnitOfWork unitOfWork)
     {
         _dreamsRepository = dreamsRepository;
         _userRepository = userRepository;
+        _unitOfWork = unitOfWork;
     }
 
     public async Task<bool> Handle(DeleteDreamCommand request, CancellationToken cancellationToken)
@@ -26,13 +28,19 @@ public class DeleteDreamCommandHandler : IRequestHandler<DeleteDreamCommand, boo
             throw new BadRequestException("Post does not exist.");
         
         var userRole = await _userRepository.GetRole(request.WhoRequested);
+
+        if (userRole == null)
+            throw new NotFoundException("You do not exist");
         
-        var ownsDream = await _dreamsRepository.UserOwnsPost(request.WhoRequested, request.DreamId) ||
-                        ModerationAccessHelper.CanEditPost(userRole);
+        var ownsDream = oldPost.Author.Id == request.WhoRequested ||
+                        ModerationAccessHelper.CanEditPost((DreamsRole)userRole);
         
         if(!ownsDream)
             throw new BadRequestException("You do not own this dream.");
         
-        return await _dreamsRepository.DeleteDreamAsync(request.DreamId);
+        await _dreamsRepository.DeleteDreamAsync(request.DreamId);
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+        return true;
     }
 }
